Modification cloud-servicechain-state.yang key
Due to elan-service-chain-state is an augmentation of
elan-instance, elan-instance-name can be removed from
elan-to-pseudo-port-data list key. The new key will be
composed with elanLportTag and scfTag.

ScfTag is now defined as an unsigned 32-bit number

+++ vpnservice/cloud-servicechain/cloud-servicechain-impl/src/test/java/org/opendaylight/netvirt/cloudservicechain/VPNServiceChainHandlerTest.java
@@ -89,7 +89,7 @@ public class VPNServiceChainHandlerTest {
     final String RD = "1.1.1.1:10";
     final String vpnName = "1";
     final long vpnId = 1;
-    final int scfTag = 1;
+    final long scfTag = 1L;
     final int servChainTag = 100;
     final int dpnId = 1;
     final int lportTag = 1;
